+++ example/lib/app/app.dart
@@ -1,8 +1,8 @@
-import 'package:chewie/chewie.dart';
-import 'package:chewie_example/app/theme.dart';
-import 'package:flutter/material.dart';
+import 'package:cached_chewie/cached_chewie.dart';
+import 'package:cached_chewie_example/app/theme.dart';
 // ignore: depend_on_referenced_packages
-import 'package:video_player/video_player.dart';
+import 'package:cached_video_player/cached_video_player.dart';
+import 'package:flutter/material.dart';
 
 class ChewieDemo extends StatefulWidget {
   const ChewieDemo({
@@ -20,8 +20,8 @@ class ChewieDemo extends StatefulWidget {
 
 class _ChewieDemoState extends State<ChewieDemo> {
   TargetPlatform? _platform;
-  late VideoPlayerController _videoPlayerController1;
-  late VideoPlayerController _videoPlayerController2;
+  late CachedVideoPlayerController _videoPlayerController1;
+  late CachedVideoPlayerController _videoPlayerController2;
   ChewieController? _chewieController;
 
   @override
@@ -45,9 +45,9 @@ class _ChewieDemoState extends State<ChewieDemo> {
 
   Future<void> initializePlayer() async {
     _videoPlayerController1 =
-        VideoPlayerController.network(srcs[currPlayIndex]);
+        CachedVideoPlayerController.network(srcs[currPlayIndex]);
     _videoPlayerController2 =
-        VideoPlayerController.network(srcs[currPlayIndex]);
+        CachedVideoPlayerController.network(srcs[currPlayIndex]);
     await Future.wait([
       _videoPlayerController1.initialize(),
       _videoPlayerController2.initialize()