+++ example/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:chewie_example/app/app.dart';
+import 'package:cached_chewie_example/app/app.dart';
 import 'package:flutter/material.dart';
 
 void main() {