+++ lib/cached_chewie.dart
@@ -1,4 +1,4 @@
-library chewie;
+library cached_chewie;
 
 export 'src/chewie_player.dart';
 export 'src/chewie_progress_colors.dart';