+++ lib/src/cupertino/cupertino_progress_bar.dart
@@ -1,8 +1,8 @@
-import 'package:chewie/src/chewie_progress_colors.dart';
-import 'package:chewie/src/progress_bar.dart';
+import 'package:cached_chewie/src/chewie_progress_colors.dart';
+import 'package:cached_chewie/src/progress_bar.dart';
+import 'package:cached_video_player/cached_video_player.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:video_player/video_player.dart';
 
 class CupertinoVideoProgressBar extends StatelessWidget {
   CupertinoVideoProgressBar(
@@ -15,7 +15,7 @@ class CupertinoVideoProgressBar extends StatelessWidget {
   })  : colors = colors ?? ChewieProgressColors(),
         super(key: key);
 
-  final VideoPlayerController controller;
+  final CachedVideoPlayerController controller;
   final ChewieProgressColors colors;
   final Function()? onDragStart;
   final Function()? onDragEnd;