+++ lib/src/material/material_controls.dart
@@ -1,18 +1,18 @@
 import 'dart:async';
 
-import 'package:chewie/src/center_play_button.dart';
-import 'package:chewie/src/chewie_player.dart';
-import 'package:chewie/src/chewie_progress_colors.dart';
-import 'package:chewie/src/helpers/utils.dart';
-import 'package:chewie/src/material/material_progress_bar.dart';
-import 'package:chewie/src/material/widgets/options_dialog.dart';
-import 'package:chewie/src/material/widgets/playback_speed_dialog.dart';
-import 'package:chewie/src/models/option_item.dart';
-import 'package:chewie/src/models/subtitle_model.dart';
-import 'package:chewie/src/notifiers/index.dart';
+import 'package:cached_chewie/src/center_play_button.dart';
+import 'package:cached_chewie/src/chewie_player.dart';
+import 'package:cached_chewie/src/chewie_progress_colors.dart';
+import 'package:cached_chewie/src/helpers/utils.dart';
+import 'package:cached_chewie/src/material/material_progress_bar.dart';
+import 'package:cached_chewie/src/material/widgets/options_dialog.dart';
+import 'package:cached_chewie/src/material/widgets/playback_speed_dialog.dart';
+import 'package:cached_chewie/src/models/option_item.dart';
+import 'package:cached_chewie/src/models/subtitle_model.dart';
+import 'package:cached_chewie/src/notifiers/index.dart';
+import 'package:cached_video_player/cached_video_player.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:video_player/video_player.dart';
 
 class MaterialControls extends StatefulWidget {
   const MaterialControls({
@@ -31,7 +31,7 @@ class MaterialControls extends StatefulWidget {
 class _MaterialControlsState extends State<MaterialControls>
     with SingleTickerProviderStateMixin {
   late PlayerNotifier notifier;
-  late VideoPlayerValue _latestValue;
+  late CachedVideoPlayerValue _latestValue;
   double? _latestVolume;
   Timer? _hideTimer;
   Timer? _initTimer;
@@ -44,7 +44,7 @@ class _MaterialControlsState extends State<MaterialControls>
   final barHeight = 48.0 * 1.5;
   final marginSize = 5.0;
 
-  late VideoPlayerController controller;
+  late CachedVideoPlayerController controller;
   ChewieController? _chewieController;
 
   // We know that _chewieController is set in didChangeDependencies
@@ -301,7 +301,7 @@ class _MaterialControlsState extends State<MaterialControls>
   }
 
   GestureDetector _buildMuteButton(
-    VideoPlayerController controller,
+    CachedVideoPlayerController controller,
   ) {
     return GestureDetector(
       onTap: () {