+++ lib/src/player_with_controls.dart
@@ -1,9 +1,9 @@
-import 'package:chewie/src/chewie_player.dart';
-import 'package:chewie/src/helpers/adaptive_controls.dart';
-import 'package:chewie/src/notifiers/index.dart';
+import 'package:cached_chewie/src/chewie_player.dart';
+import 'package:cached_chewie/src/helpers/adaptive_controls.dart';
+import 'package:cached_chewie/src/notifiers/index.dart';
+import 'package:cached_video_player/cached_video_player.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:video_player/video_player.dart';
 
 class PlayerWithControls extends StatelessWidget {
   const PlayerWithControls({Key? key}) : super(key: key);
@@ -45,7 +45,8 @@ class PlayerWithControls extends StatelessWidget {
               child: AspectRatio(
                 aspectRatio: chewieController.aspectRatio ??
                     chewieController.videoPlayerController.value.aspectRatio,
-                child: VideoPlayer(chewieController.videoPlayerController),
+                child:
+                    CachedVideoPlayer(chewieController.videoPlayerController),
               ),
             ),
           ),