+++ CHANGELOG.md
@@ -253,7 +253,7 @@
 
 Initial version of Chewie, the video player with a heart of gold.
 
-  * Hand a VideoPlayerController to Chewie, and let it do the rest.
+  * Hand a CachedVideoPlayerController to Chewie, and let it do the rest.
   * Includes Material Player Controls
   * Includes Cupertino Player Controls
   * Spike version: Focus on good looking UI. Internal code is sloppy, needs a refactor and tests